Cambios varios: Configuration de los services y la Api Gateway hecha, dtos, services y repositories del microservicio prueba agregados (en su mayoria hechos), proyecto del service prueba correctamente implementado

+++ prueba-service/src/main/java/com/utn/frc/backend/pruebaservice/controllers/PruebaController.java
@@ -28,7 +28,7 @@ public class PruebaController {
     }
 
     @PostMapping("/{id}/finalizar")
-    public ResponseEntity<String> finalizarPrueba(@PathVariable Long id, @RequestBody PruebaDTO pruebaDTO) {
+    public ResponseEntity<String> finalizarPrueba(@PathVariable Integer id, @RequestBody PruebaDTO pruebaDTO) {
         pruebaService.finalizarPrueba(id, pruebaDTO);
         return ResponseEntity.ok("Prueba finalizada exitosamente.");
     }
